+++ .eslintcache
@@ -1,1 +1,1 @@
-[{"D:\\Tutotials\\ReactJS\\udemy\\react-complete-guide\\src\\index.js":"1","D:\\Tutotials\\ReactJS\\udemy\\react-complete-guide\\src\\App.js":"2","D:\\Tutotials\\ReactJS\\udemy\\react-complete-guide\\src\\components\\ExpenseItem.js":"3","D:\\Tutotials\\ReactJS\\udemy\\react-complete-guide\\src\\components\\ExpenseDate.js":"4","D:\\Tutotials\\ReactJS\\udemy\\react-complete-guide\\src\\components\\Expenses.js":"5","D:\\Tutotials\\ReactJS\\udemy\\react-complete-guide\\src\\components\\Card.js":"6"},{"size":142,"mtime":1616592305000,"results":"7","hashOfConfig":"8"},{"size":651,"mtime":1643418431819,"results":"9","hashOfConfig":"8"},{"size":473,"mtime":1643433382374,"results":"10","hashOfConfig":"8"},{"size":519,"mtime":1643414844835,"results":"11","hashOfConfig":"8"},{"size":794,"mtime":1643433396008,"results":"12","hashOfConfig":"8"},{"size":182,"mtime":1643433856088,"results":"13","hashOfConfig":"8"},{"filePath":"14","messages":"15","errorCount":0,"fatalErrorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"16"},"3t4mzk",{"filePath":"17","messages":"18","errorCount":0,"fatalErrorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"19","messages":"20","errorCount":0,"fatalErrorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"21","messages":"22","errorCount":0,"fatalErrorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"23","messages":"24","errorCount":0,"fatalErrorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"25","messages":"26","errorCount":0,"fatalErrorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"D:\\Tutotials\\ReactJS\\udemy\\react-complete-guide\\src\\index.js",[],["27","28"],"D:\\Tutotials\\ReactJS\\udemy\\react-complete-guide\\src\\App.js",[],"D:\\Tutotials\\ReactJS\\udemy\\react-complete-guide\\src\\components\\ExpenseItem.js",[],"D:\\Tutotials\\ReactJS\\udemy\\react-complete-guide\\src\\components\\ExpenseDate.js",[],"D:\\Tutotials\\ReactJS\\udemy\\react-complete-guide\\src\\components\\Expenses.js",[],"D:\\Tutotials\\ReactJS\\udemy\\react-complete-guide\\src\\components\\Card.js",[],{"ruleId":"29","replacedBy":"30"},{"ruleId":"31","replacedBy":"32"},"no-native-reassign",["33"],"no-negated-in-lhs",["34"],"no-global-assign","no-unsafe-negation"]
+[{"D:\\Tutotials\\ReactJS\\udemy\\react-complete-guide\\src\\index.js":"1","D:\\Tutotials\\ReactJS\\udemy\\react-complete-guide\\src\\App.js":"2","D:\\Tutotials\\ReactJS\\udemy\\react-complete-guide\\src\\components\\Expenses\\Expenses.js":"3","D:\\Tutotials\\ReactJS\\udemy\\react-complete-guide\\src\\components\\Expenses\\ExpenseItem.js":"4","D:\\Tutotials\\ReactJS\\udemy\\react-complete-guide\\src\\components\\UI\\Card.js":"5","D:\\Tutotials\\ReactJS\\udemy\\react-complete-guide\\src\\components\\Expenses\\ExpenseDate.js":"6"},{"size":142,"mtime":1616592305000,"results":"7","hashOfConfig":"8"},{"size":660,"mtime":1643499510018,"results":"9","hashOfConfig":"8"},{"size":798,"mtime":1643499426390,"results":"10","hashOfConfig":"8"},{"size":477,"mtime":1643499417101,"results":"11","hashOfConfig":"8"},{"size":182,"mtime":1643499382094,"results":"12","hashOfConfig":"8"},{"size":519,"mtime":1643499390069,"results":"13","hashOfConfig":"8"},{"filePath":"14","messages":"15","errorCount":0,"fatalErrorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"16"},"3t4mzk",{"filePath":"17","messages":"18","errorCount":0,"fatalErrorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"19","messages":"20","errorCount":0,"fatalErrorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"21","messages":"22","errorCount":0,"fatalErrorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"23","messages":"24","errorCount":0,"fatalErrorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"25","messages":"26","errorCount":0,"fatalErrorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"D:\\Tutotials\\ReactJS\\udemy\\react-complete-guide\\src\\index.js",[],["27","28"],"D:\\Tutotials\\ReactJS\\udemy\\react-complete-guide\\src\\App.js",[],"D:\\Tutotials\\ReactJS\\udemy\\react-complete-guide\\src\\components\\Expenses\\Expenses.js",[],"D:\\Tutotials\\ReactJS\\udemy\\react-complete-guide\\src\\components\\Expenses\\ExpenseItem.js",[],"D:\\Tutotials\\ReactJS\\udemy\\react-complete-guide\\src\\components\\UI\\Card.js",[],"D:\\Tutotials\\ReactJS\\udemy\\react-complete-guide\\src\\components\\Expenses\\ExpenseDate.js",[],{"ruleId":"29","replacedBy":"30"},{"ruleId":"31","replacedBy":"32"},"no-native-reassign",["33"],"no-negated-in-lhs",["34"],"no-global-assign","no-unsafe-negation"]
